Some more events and custom actions

+++ app/src/main/java/be/ac/ulg/montefiore/group03/agilegame/gamelogic/Events/Feature_Event.java
@@ -22,7 +22,7 @@ public class Feature_Event extends Event{
         this.depend = f;
     }
 
-    public Feature_Event(int id, Date d, Features f, Skills s, int delay ){
+    public Feature_Event(int id, Date d, Features f, Skills s, double delay ){
         super(id,d,s,delay);
         this.depend = f;
     }
